PM: Execute PM suspend/resume callbacks in a task
A new PM task has been added in a regular SYS/BIOS environment
to execute the various user registered suspend & resume callback
functions. This allows the users to use semaphores in the callback
functions.

The new task is not required for SMP/BIOS because the SMP/BIOS
already executes all the Power Suspend functionality within a
dedicated task on each of the IPU cores. The BIOS Power_suspend
calls the callback functions in these task contexts.

NOTE: The callback functions are ONLY called by CORE0 in SMP/BIOS.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>