IPC: Initialize procId variables in interrupt send functions
The procId variables are static on a particular base image and
are used in the InterruptXXX_intSend functions to eliminate the
need to do a function lookup for the id every time a message
needs to be sent. The variables are currently initialized in the
InterruptXXX_intRegister function. The xxx_intRegister function
is used for enabling a Rx interrupt to process received messages,
and should be orthogonal to the xxx_intSend function, which is
solely used for sending a message.

The procId initialization has therefore been moved to the
xxx_intSend function itself to break the above dependency. This
ensures that the variables are always initialized, and would
avoid any exceptions when sending a message even before a Rx
interrupt has been registered.

This behavior has been observed in the case of an early crash.
The VirtQueue_postCrashToMailbox tries to notify MPU in case
of an exception during the initial boot and this can lead to
another secondary crash due to uninitialized procId variables.
This nested-exception behavior masks the register and stack
dump of the very first crash.

Signed-off-by: Artur Troian <x0164361@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>