DEH: Disallow timer register accesses if watchdog disabled
The Watchdog feature can be enabled or disabled through a build time
configuration on the host-side. The new Watchdog implementation was
allowing transactions on the timer without really checking if the
associated timer is enabled or not. The watchdog timer registers are
accessed during scheduler switches and without checking for the
timer statys can cause a continuous exception on the slave processors
when watchdog feature is disabled.

The previous implementation used a valid timer base address itself
as the check, but the current implementation always assigns the base
address for the timer used.  An additional status mode has therefore
been added for each timer and checked before programming the timer
registers.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Subramaniam C.A <subramaniam.ca@ti.com>