DEH: Restore watchdog timer context to support Device OFF mode
During OFF mode suspend, the states of most IP blocks in the SOC
are lost when the IP is powered off. Without software action to
restore the lost state, the modules will be left in their reset
states and will not be able to resume where they left off.

The timers will need to be reinitialized after Device OFF. A new
Watchdog_restore function is registered as a resume callback hook
with the IpcPower module and reinitializes the watchdog timer by
reprogramming the timer registers. The timer registers are
reconfigured with the same set of values as during the initial
boot except for the counter register, which is programmed with
the regular watchdog timeout value.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>