Build: Clear internal configuration settings during SMP config
The build for SMP-mode IPU target is enabled by configuring the
build system to an exclusive SMP build mode using 'make smp_config'.

An internal configuration file is created in the above step, and the
entries in this file can be repeated for consecutive 'make smp_config'
invocations. The 'unconfig' target is therefore added to 'smp_config'
make target to remove the internal file.

Signed-off-by: Suman Anna <s-anna@ti.com>