Build: Update to CGT 7.2.9 compiler for DSP
The compiler for DSP is updated to the latest version in the 7.2
release stream, 7.2.9. Only the README is updated as the compiler
version used is solely dictated by the C6000CGTOOLPATH environment
variable value.

The newer version is recommended to be used so that various compiler
fixes can be picked up. Please look up "DefectHistory.txt" file in
the compiler installation directory for a summary of fixes.

Signed-off-by: Suman Anna <s-anna@ti.com>