Utils: Add a new Version module
A new module, Version, is added to retrieve and present the
version and tools information associated with the ELF base
images. The information is retrieved during link phase and
stored in a non-loadable section named ".version".

The version information comprises of the git tag, branch and
top most commit information (only when the image is built
from source code stored on a git repository).

The information is retrieved automatically when the module
is included in the base-image configuration file.
    var Version = xdc.useModule('ti.utils.Version');

The version information can be read using the standard ELF
utility, as
    readelf -p .version <base_image>

The version information can also be completely overridden by
the user to add his own generated details.
  Eg:
    Version.custom = "My custom version";

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>