HDMIWA: Publish custom tick function to IpcPower module
The IpcPower module needs a handle to the BIOS Tick timer
handle inorder to restore its context. The IpcPower module
retrieves this handle by searching for standard clock tick
functions among the configured timers on the target.

The HDMIWA uses a custom tick function and internally calls
the Clock_tick function, and so needs to publish this custom
function to the IpcPower module. This will allow the IpcPower
module to cross-check against this custom function as well
in finding the tick timer handle.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>