ResMgr: Add new API to request resource data
A new API, IpcResource_requestData, is added to request data
associated with a particular resource.

This API currently supports to request data about the maximum
frequency on a processor device. The checking for the proper
device type is performed on the host-side.

The interface is as follows:
  typedef enum {
      IpcResource_ReqDataType_MAX_FREQ
  } IpcResource_ReqDataType;

  Int IpcResource_requestData(IpcResource_Handle handle,
                              IpcResource_ResHandle resHandle,
                              IpcResource_ReqDataType dataType,
                              Void *data);

User is supposed to pass in a structure large enough (based on
dataType) to get back the requested information.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>