Build: Update the default installation folder
The default installation location for all the new tools
is /opt/ti (installed as root). The build system has been
updated to reflect this location as the default value for
the BIOSTOOLSROOT variable. This environment variable need
not be defined if all tools are installed using the default
location.

The only exception at present is the ARM Code Generation
Tools whose installation location is /opt/TI. This has been
corrected in newer versions of the ARM CGT, but for now
requires the installation location to be changed manually
to /opt/ti.

Signed-off-by: Suman Anna <s-anna@ti.com>