Configs: Skip timer frequency check configuration
The upgrade to new BIOS 6.33.06.50 adds a check for the timer
frequency by measuring it using a TimeStamp module whose own
timer is derived from the processor clock. The processor may
not always be clocked at OPP100, and so may be at a different
frequency than nominal one. The timer check for the BIOS tick
timer fails in this case, and throws an abort right during the
initial image boot.

This is avoided by programming a configuration option with the
dmTimer module to skip this check, and thereby have the image
booted always just as with the previous BIOS version.

Signed-off-by: Suman Anna <s-anna@ti.com>