Configs: Configure default heap size using XDC Memory module
The default heap size is defined using the BIOS.heapSize property
from BIOS module. This is changed to use the Memory.defaultHeapSize
property from the XDC Memory module.

This change removes any chance for a warning thrown by BIOS code
if both properties are set, or if a default Memory heap instance
is defined. The BIOS module is automatically configured to use
the Memory.defaultHeapSize internally, so this change is backward
compatible and removes possible warning messages due to external
default heap configurations.

Signed-off-by: Suman Anna <s-anna@ti.com>