Gates: Add support to track hwspinlock locking
HwSpinlocks are used to arbitrate access to a common resource
between two or more processors. HwSpinlocks could be left in
a locked state if any device exception happens on the processor
holding the lock. Such a scenario calls for a way to recover /
reset the locked HwSpinlocks so that the arbitration can be
handled properly between the processors again.

This patch introduces a shared state that can be exposed to
the host via a resource table entry. The shared state for the
spinlocks is basically a bitfield (1 bit per hwspinlock) that
would be set each time the corresponding hwspinlock is held
and reset when released. The size of the bit field array is
also exported to aid the host in indexing into the array. The
host can inspect this shared state and reset the spinlocks
held by the remote processor while reloading the remote
processor.

Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>