Resources: Add hwspinlock state as a custom resource
HwSpinlocks can be left locked in the event of a device
exception. To gracefully exit such a scenario, the state
of the hwspinlocks (stored in a bitfield array) is exposed
as a custom resource table entry. The number of the locks
is also exposed. This information can be used by the host
to reset the locks and recover the remote processor.

Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>