Utils: Fix for use of double quotes in version commit message
The version info generation adds the top-most commit message of
the sysbios-rpmsg tree and the tree building the base-image. The
use of double quotes in the commit message (like in a revert or
a merge commit) was not handled properly when constructing the
version string. The quote character was seen by the compiler as
the end of the string instead of another valid character, and is
resulting in a build error at linking stage.

The version string generation for the commit message is corrected
to handle the presence of the quotes.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>