Utils: Add a new genextelf utility for pre-signing images
A new utility, called "genextelf", is added that processes an IPU
ELF image and gives out an extended pre-signed ELF image. The ELF
image outputed would be fed into a signing tool to generate the
final signed IPU image.

The utility adds new sections to the image with the various data
needed for the signing process such as MMU entries, the TOC data
etc. The genextelf utility is similar to the genextbin utility,
which was used for signing the binary files.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>