Configs: Move Power module from IpcCommon to specific cfg files
The BIOS Power module for SMP-mode has a different path from the
regular SYS/BIOS Power module path in the new unified BIOS 6.34
release stream.

The inclusion of the module is therefore moved from the generic
IpcCommon.cfg.xs files to each of the specific core config files,
with the proper module included.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>