Build: Upgrade the SMP build to use BIOS 6.34.01.14
SMP/BIOS functionality is merged with the regular SYS/BIOS into
a single product in BIOS 6.34. SMP-mode can be specifically
built for by setting the BIOS.smpEnabled configuration property
by the application.

The build system is updated to use the current latest SMP-mode
BIOS version, BIOS 6.34.01.14. This is the recommended version to
be used for SMP/BIOS to maintain functional feature compatibility
with the previous SMP/BIOS 1.00.00.28_eng version. The new version
needs the applications to replace some of the APIs used in
SMP-mode.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>