Build: Update to newer BIOS 6.34.01.14 version
The build system is updated to use the latest BIOS release,
BIOS 6.34.01.14. This newer version is the latest release in
the BIOS 6.34 release stream that unifies the SMP/BIOS and
SYS/BIOS into a single BIOS product package.

The distinction is made with the BIOS.smpEnabled configuration
parameter within BIOS. Applications would have to specifically
configure this to build for SMP.

The newer BIOS version is built with ARM CGT 5.0.1, and requires
previous existing libraries to be rebuilt against the newer
version using the new compiler to reduce any portability issues.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>