Resources: Define memory region type for carveout entries
Each memory resource entry in the resource table can optionally
define a memory region type that provides information on the
type of memory access/usage made by the remote processor on that
 memory region. This is defined using an existing reserved field
in the structures defining the carveout and devmem resources.

This memory region type is used for determining the type of
firewalls to be installed on the regions during secure playback
usecase.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>