Utils: Update suspend address lookup for SMP-mode with BIOS 6.34
The suspend address for ARM M3/M4 targets is currently looked up using
the symbol name, ti_sysbios_family_arm_ducati_omap4430_readyIdleCore0
on Core0. This symbol name is no longer valid in SMP-mode with BIOS
6.34 release stream, as the SMP-mode Power module is different and has
a different package path from the non-SMP Power module.

The symbol address, ti_sysbios_family_arm_ducati_smp_readyIdleCore0,
is used to detect the suspend address for a M3/M4 base-image file.
This symbol is used only when one base-image is supplied to the
genrprc utility, as is the case for SMP.

Signed-off-by: Suman Anna <s-anna@ti.com>