Revert "PM: Enable mbox interrupt in wugen for DSP before suspending"
This reverts commit babf5fe293ef5ff912c198c905cb0b07f6e0df85.

The above commit is added as a means to have the DSP resumed properly
after the remote processor has been suspended from the host side. The
root-cause for the original issue is unknown, but the above solution
is no longer needed with the latest host-side code.

The remote processor would be put into reset and the mailbox interrupt
register context is saved during a suspend. The DSP Wakeup Generator
would not need any user-configured interrupts to be enabled before the
DSP is suspended for it to wake up properly. This revert changes the
code back to this expected behavior.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>