PM: Enable timer register restoration for non-SMP targets
The Timer register restoration API was supported and made
available first only in the SMP/BIOS product. This API is
available for both SMP and non-SMP modes in the unified BIOS
6.34 release (starting from 6.34.01.14).

The restore functionality for non-SMP targets was previously
commented out, and relied on the timers being restored by the
host-side code. The Timers are now properly restored even for
non-SMP targets and truely enables any resume from Device OFF
for processors running regular SYS/BIOS, and removes the
dependency with the host-side restore functionality.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>