DEH: Improve the detection logic for OMAP5 ES 2.x
Additional logic has been added to make the detection logic
of OMAP5 ES 2.x silicon robust. The additional logic relies
on checking the debug id version of the MPU (Cortex A15)
processor core through the emulation domain. This additional
logic helps on some earlier samples where the silicon version
id is not reflected properly.

NOTE: This patch is a temporary workaround, it breaks OMAP5
ES1.0 and needs to be reverted.

Signed-off-by: Suman Anna <s-anna@ti.com>