DEH: Adjust for OMAP5 ES 2.x PRCM address changes for IPU
The Watchdog module in IPU relies on two GPTimers, GPT9 & GPT11.
These timers belong to the L4_PER clock domain. The L4_PER clock
domain has been moved into the CORE power domain in OMAP5 ES2.x,
and therefore has the respective PRCM addresses changed.

The Watchdog module relies on the clock status of the timers to
dictate enabling or disabling of the watchdog timers. Logic has
been added to detect the OMAP SoC silicon revision id and adjust
the addresses accordingly. This change is not needed on DSP since
it uses a GPTimer internal to the ABE domain whose addresses
remain the same.

Signed-off-by: Suman Anna <s-anna@ti.com>