Revert "Build: Remove building for core1 platforms"
This reverts commit 17614baf51e89874f750e672495a0d2ccb33d392.

The IPU Core1 platforms have been added back to the possible
target platform list in order to bring back the support for
booting the two IPU cores in a non-SMP mode.

The latest host-side software expects a single ELF image for
the IPU cores, so this will be achieved by developing a new
utility that will combine the two ELF base images into a
suitable single ELF image.

Signed-off-by: Suman Anna <s-anna@ti.com>