Utils: Add a new gencmbelf utility to combine two ELF images
A new utility, called "gencmbelf", is added that combines two non-SMP
IPU ELF images into a suitable semi-standard single ELF image to be
used instead of a single SMP ELF image.

The utility requires that the input images are completely stripped of
any debug or symbol data info, as the utility is currently not designed
to combine these data. The generated output file is an ELF image and
can be passed to the readelf utility to retrieve information just like
any other ELF file.

This utility is designed for the current ELF images built with or for
sysbios-rpmsg, and not a generic utility.

Signed-off-by: Suman Anna <s-anna@ti.com>