Resources: Add trace1 and vdev for AppM3 in non-SMP
This patch adds the additional trace and virtio device entries
into the resource table, necessary to enable the IPU Core1
processor to be booted in a non-SMP mode.

The trace entry value for Core1 is currently hard-coded and is
expected to be fixed up in the gencmbelf utility, to support a
dynamic trace address for Core1.

Signed-off-by: Suman Anna <s-anna@ti.com>