IPC: Redefine VirtQueue ids for IPU Core1 to support non-SMP
The indices for IPU Core1/Appm3 have been moved far apart in a
previous patch,
  c2a1808 "IPC: Redefine VirtQueue ids sequentially for each processor"
as it was expected to be used only in a SMP-mode.

These indices have been changed back to their older values of 2 and
3 so that the Core1 can be run alongside Core0 in a non-SMP mode.
These values differ from the current philosophy of starting the
indices from 0 for each processor. These values are chosen since
Core1 will still be booted together with the Core0 processor - the
two processors are treated as a single entity by the host-side
software.

Signed-off-by: Suman Anna <s-anna@ti.com>