Build: Generate default ELF binaries for both SMP & non-SMP modes
The build is updated to automatically generate the desired output
ELF binaries for both SMP and non-SMP modes for OMAP4 & OMAP5 platforms.
The SMP-mode and non-SMP DSP binaries are simply symbol and debug info
stripped versions of the actual ELF base-images. The non-SMP IPU binaries
use the genrprc script which internally invokes the gencmbelf utility
to generate a single combined ELF image.

Signed-off-by: Suman Anna <s-anna@ti.com>