Utils: Update genrprc utility to invoke gencmbelf
The genrprc utility has been updated to invoke gencmbelf
utility to generate a combined ELF image, using stripped
versions of the passed in ELF images, but only after
retrieving the needed symbols from the input files. The
syntax is same as the previous version, but all the input
and output files are expected to the ELF M3 files.

These symbols are retrieved from Core1 image file and
used for updating the associated entries in the resource
table which is present only in Core0 image. Currently,
only the symbols needed for the Core1 trace entry are
retrieved.

Signed-off-by: Suman Anna <s-anna@ti.com>