Utils: Update genrprc utility to pass additional arguments
The genrprc utility has been updated to invoke gencmbelf
utility with additional arguments pertinent to Core1 image
for properly updating the associated entries in the resource
table. The arguments are generated by using standard ELF
utilities and looking up specific symbols/sections.

The additional arguments include only the trace entries
(address & size for each) as that is the only changeable
resource for Core1 present in the resource table currently.

Signed-off-by: Suman Anna <s-anna@ti.com>