Resources: Move IPC data section just after the vring
The physical memory for the carveout resource entries are
allocated in the order they are present in the resource table.
The IPC data carveout entry is currently after the text and
data entries, and is moved ahead of them so that it's physical
memory is allocated next to the IPC vring area. This will allow
the IPC regions to be better firewalled together for secure
playback usecase.

Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>