Utils: Fix the ELF format corruption in genextelf
The genextelf utility is ignoring the alignment of the different
ELF sections, and as such is outputting a ELF image with corrupted
ELF sections. The IPU processor booted successfully even with the
invalid ELF format when using a SMP ELF image, but does not work
with the combined non-SMP ELF image.

The genextelf utility has been updated and the logic reworked to
adhere to the ELF section and format restrictions. The section
headers for the newly added sections are also ordered correctly
as per their corresponding section data/offsets.

Signed-off-by: Suman Anna <s-anna@ti.com>