Utils: Fix the Core1 section offsets in gencmbelf
The current logic in gencmbelf utility for computing the section
offsets of Core1 image is depending on the original section offsets,
which also include the ELF header size from the second image. The
combined ELF image will only have a single ELF header, and so the
Core1 section offset calculation need to account for this.

TODO: The alignment of different program sections in Core1 will also
affect the section offsets in the combined image, this still needs to
be addressed. It does not affect the processor loading or booting
though.

Signed-off-by: Suman Anna <s-anna@ti.com>