Utils: Fix the MMU table TOC entries in genextelf
The computation of offsets and physical addresses for the
MMU table static entries is off by 4, causing issues with
the authentication. This logic has been corrected.

The memregions for the MMU table TOC entries are also
populated, and the last TOC entry is given an unknown
memory region, instead of the previous count.

Signed-off-by: Suman Anna <s-anna@ti.com>